Extract definition of the Weather view controller's View
